Move helper methods to the abstract intent
